Share dso struct between native and JS code. NFC
This avoids passing the filename as an additional
paramater to JS and is a pre-cursor to having the JS side
communicate more information about the loaded DSO back
to the native side.
