+++ src/library_dylink.js
@@ -763,11 +763,11 @@ var LibraryDylink = {
 
   // void* dlopen(const char* filename, int flags);
   $dlopenInternal__deps: ['$FS', '$ENV', '$dlSetError'],
-  $dlopenInternal: function(filenameAddr, flags, jsflags, handle) {
+  $dlopenInternal: function(handle, flags, jsflags) {
     // void *dlopen(const char *file, int mode);
     // http://pubs.opengroup.org/onlinepubs/009695399/functions/dlopen.html
     var searchpaths = [];
-    var filename = UTF8ToString(filenameAddr);
+    var filename = UTF8ToString(handle + {{{ C_STRUCTS.dso.name }}});
 
     var isValidFile = function(filename) {
       var target = FS.findObject(filename);
@@ -817,14 +817,14 @@ var LibraryDylink = {
 
   _dlopen_js__deps: ['$dlopenInternal'],
   _dlopen_js__sig: 'iiii',
-  _dlopen_js: function(filename, flags, handle) {
+  _dlopen_js: function(handle, flags) {
 #if ASYNCIFY
     return Asyncify.handleSleep(function(wakeUp) {
       var jsflags = {
         loadAsync: true,
         fs: FS, // load libraries from provided filesystem
       }
-      var promise = dlopenInternal(filename, flags, jsflags, handle);
+      var promise = dlopenInternal(handle, flags, jsflags);
       promise.then(wakeUp).catch(function() { wakeUp(0) });
     });
 #else
@@ -832,7 +832,7 @@ var LibraryDylink = {
       loadAsync: false,
       fs: FS, // load libraries from provided filesystem
     }
-    return dlopenInternal(filename, flags, jsflags, handle);
+    return dlopenInternal(handle, flags, jsflags);
 #endif
   },
 
@@ -844,9 +844,10 @@ var LibraryDylink = {
 #endif
   ],
   _emscripten_dlopen_js__sig: 'viiiii',
-  _emscripten_dlopen_js: function(filename, flags, handle, onsuccess, onerror) {
+  _emscripten_dlopen_js: function(handle, flags, onsuccess, onerror) {
     function errorCallback(e) {
-      dlSetError('Could not load dynamic lib: ' + UTF8ToString(filename) + '\n' + e);
+      var filename = UTF8ToString({{{ makeGetValue('handle', C_STRUCTS.dso.name, '*') }}});
+      dlSetError('Could not load dynamic lib: ' + filename + '\n' + e);
       {{{ runtimeKeepalivePop() }}}
       callUserCallback(function () { {{{ makeDynCall('vi', 'onerror') }}}(handle); });
     }
@@ -856,7 +857,7 @@ var LibraryDylink = {
     }
 
     {{{ runtimeKeepalivePush() }}}
-    var promise = dlopenInternal(filename, flags, { loadAsync: true }, handle);
+    var promise = dlopenInternal(handle, flags, { loadAsync: true });
     if (promise) {
       promise.then(successCallback, errorCallback);
     } else {
